+++ Adj/Algebra/Morphism.hs
@@ -1,3 +1,3 @@
 module Adj.Algebra.Morphism (module Exports) where
 
-import Adj.Algebra.Morphism.Straight as Exports
+import Adj.Algebra.Morphism.Flat as Exports