+++ Adj/Algebra/Morphism/Flat.hs
@@ -1,0 +1,3 @@
+module Adj.Algebra.Morphism.Flat where
+
+newtype Flat morphism source target = Flat (morphism source target)